% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor2.R
\docType{data}
\name{PA_time}
\alias{PA_time}
\title{Measurement times of self-reported positive affect}
\format{A numeric vector of length 6.}
\source{
Mutak, A., Pavlović, M. & Zibar, K. (2017, May). \emph{Postoje li razlike između introverata i ekstraverata u cirkadijurnim ritmovima raspoloženja?} [\emph{Are There Differences Between Introverts and Extraverts in Circadian Mood Rhythms?}]. Study presented at the 3rd \emph{Regionalni susret studenata psihologije - SPIRI} [\emph{Regional Meeting of Psychology Students - SPIRI}] conference, Rijeka, Croatia.
}
\usage{
PA_time
}
\description{
A dataset containing the measurement times (hours) of self reported positive affect contained in the data frames \code{PA_extravers} and \code{PA_introverts}.
}
\keyword{datasets}
