% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor2.R
\docType{data}
\name{PANAS_november}
\alias{PANAS_november}
\title{Self-reported mood}
\format{A data frame with 30 rows and 20 variables:
\describe{
  \item{Time}{Day of measurement.}
  \item{Subject 1, Subject 2, Subject 3, Subject 4, Subject 5, Subject 6, Subject 7, Subject 8, Subject 9, Subject 10, Subject 11, Subject 12, Subject 13, Subject 14, Subject 15, Subject 16, Subject 17, Subject 18, Subject 19}{Results of 19 subjects on the PANAS questionnaire.}
  }}
\source{
Mutak, A. (2016, May). \emph{Kako struktura tjedna utječe na naše raspoloženje?} [\emph{How Does the Structure of the Week Impact Our Mood?}]. Study presented at the 5th \emph{Psihozij} conference, Zagreb, Croatia.
}
\usage{
PANAS_november
}
\description{
A dataset containing the responses of 19 subjects on the shortened version of the PANAS questionnaire (Watson, Clark & Tellegen, 1988) in November 2015.
}
\details{
Measurements were taken every day after 8 PM.
}
\note{
The data contained in this dataset has been reduced compared to the original data that included more subjects. This dataset contains only the subjects that have responded to the PANAS questionnaire on more than 85\% of the timepoints in both of the research cycles (July and November).
}
\references{
Watson, D., Clark, L. A. & Tellegen, A. (1988). Development and Validation of Brief Measures of Positive and Negative Affect: The PANAS Scales. \emph{Journal of Personality and Social Psychology}, \emph{54(6)}, 1063-1070.
}
\keyword{datasets}
