% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor2.R
\docType{data}
\name{PANAS_time}
\alias{PANAS_time}
\title{Measurement times of self-reported mood}
\format{A numeric vector of length 30.}
\source{
Mutak, A. i Vukasović Hlupić, T. (2017). Exogeneity of the Circaseptan Mood Rhythm and Its Relation to the Working Week. \emph{Review of Psychology}, \emph{24} (1-2), 15-28.
}
\usage{
PANAS_time
}
\description{
A dataset containing the measurement times (dates) of self reported mood contained in the data frame \code{PANAS_november}
}
\keyword{datasets}
