% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor2.R
\docType{data}
\name{temperature_zg}
\alias{temperature_zg}
\title{Daily air temperature in Zagreb}
\format{A data frame with two variables:
 \describe{
  \item{Temperature}{Air temperature in Zagreb on July 1, 2015.}
  \item{Time}{Time of the day when the temperature was measured.}
  }}
\source{
Croatian Meteorological and Hydrological Service, \url{http://www.meteo.hr}
}
\usage{
temperature_zg
}
\description{
A dataset containing the air temperature in Zagreb on July 1, 2015.
}
\details{
Measurements were taken every 30 minutes throughout the whole day.
}
\keyword{datasets}
